requirements.txt
Flask==3.1.2
pywhatkit==5.4
Pillow==11.0.0
opencv-python==4.12.0.88
img2pdf==0.6.3
reportlab==4.0.0
python-dotenv==1.0.0
werkzeug==3.1.2
pytesseract==0.3.10
